requirements.txt
mock
twisted
autobahn
klein
PyYAML
RPi.GPIO
